hemera-hzdr: remove project definition
On hemera we have no projects like on other system.
We have access groups for a few queues.
Since the slurm option `-A` is queue spezific and not user specific (will be never changed by the user) it makes no sense to have it changeable.
The current implementation also bracks the option `tbg -t anytemplate.tpl` because it was not possible to switch the queue via explicit selection of the used template.