Added absolute path determination to TBG.
Modified 'initCall' in tbg to include the script with an
absolute path instead of however it is called.
Intent being to facilitate the connection of an experiment directory and
its associated binaries.

Added the change of directory to init path to facilitate use of relative paths
for the batch script templates and configuration files.