Correct jupyter widget example
The code snippet used to illustrate the usage of the phase space visualization widget has to be corrected. `run_dir_options` expects a list of tuples of the type `( <description>, <path to data>)`.
Additionally add an example illustrating how to add additional phase space slices and to set different particle species for visualization.