More 'intuitive' nomenclature in particle shapes.
Introduced a more intuitive and slightly more consistent nomenclature of
private variables in the particle shape sources.

Re 'intuitive':
Boolean variables indicating that the argument lies **within** an interval
symmetric about 0 are named `isWithin_a_b` where `a` is the number before
the decimal point and `b` is the one after. So `isWithin_0_5` being `true`
signifies here that -0.5<= x <= 0.5 .

Re 'consistent': Just minor typos and introducing `cube` to signify a^3.

Changed `valueInSupport` to `valueOnSupport`. A point may be in the support of a function
but the function will have a value only on its support.

Changed the namespace names as well as function names to adhere to the camelCase naming convention.
Similarly changed `abs_x` to `xAbs` since `absX` looks weird and could potentially
lead to confusion with `_X`.