More 'intuitive' nomenclature in particle shapes.
Introduced a more intuitive and slightly more consistent nomenclature of
private variables in the particle shape sources.

Re 'intuitive':
Boolean variables indicating that the argument lies **within** an interval
symmetric about 0 are named `isWithin_a_b` where `a` is the number before
the decimal point and `b` is the one after. So `isWithin_0_5` being `true`
signifies here that -0.5<= x <= 0.5 .

Re 'consistent': Just minor typos and introducing `cube` to signify a^3.