Formatting corrections of the documentation.
Corrected the Python section of the plugin page to implement the
feature enumeration as an offset enumeration rather than for
it to appear in the text bulk.

Modified an equation in the PIC algorithm description.
The `cases` environment in TeX is used to format a list of cases,
such as the one used for the definition of S^0. Also it is rather uncommon
to introduce an `if` in the description, although `else` is commonly
used to describe the complement of all previous conditions.