camelCase namespace and function names.
Changed `valueInSupport` to `valueOnSupport`. A point may be in the support of a function
but the function will have a value only on its support.

Changed the namespace names as well as function names to adhere to the camelCase naming convention.
Similarly changed `abs_x` to `xAbs` since `absX` looks weird and could potentially
lead to confusion with `_X`.