fix KHI for non CUDA devices
If KHI is compiled for non CUDA devices and the field solver `DirSplitting` is enabled the compile will fail.
`DirSplitting` requires CUDA. One case defined in `cmakeFlags` is using
directional splitting.
This PR added a precompiler guard section to switch to the field solver
Yee if we compile for non CUDA devices.