Decouple particle shape assignment function order from its support
Introduce ::assignmentFunctionOrder constant to all shape structs.
openPMD-flavor output now uses this value for the particleShape attribute.
This fixes an issue with the Counter shape having zero support (for technical
reasons) and previously -1 order as the result.
Now it is also not necessary to inherit shapes from assignment functions,
remove this inheritance