Improve the initial prototype of Debye length reporting
Now the evaluation is performed for the electron species in case there is a single candidate for it.
The evaluation now averagizes over the global domain using weighting of macroparticles.
It helps to alleviate the approximate nature of the estimation.
The current version no longer judges the correctness of a setup, but merely reports the estimated temperature,
Debye length value and how many cells resolve it.