Fix the scheme of updating convolutional B in PML by half time step
The previously used scheme mistakenly assumed that updating by dt/2 twice is equivalent to updating by dt.
It resulted in worse than expected absorption.
The new scheme properly performs update of B for convolutional components.
The changes are fully contained inside the YeePML field solver and so the rest of the code is not affected.