expose compute current worker multiplier
Allow the user to tune the number of worker/threads uses to compute the
current.
The worker multiplier is now part of the strategy used to configure the
compute current task and kernel behavior.

The default for CUDA devices is changed from `2` to `1` with this PR.
This will result in a better prformance for all devices with compute
architecture 6.0+. On kepler the performance will be slightly reduced.