Remove unnecessary FieldBackground::fillSimulation() and make refactoring
This method turned uot unnecessary, both for the old and duplicateField modes.
Its results are always overwritten by the following field exchange, and this is the correct logic.

Removing the method enabled simplifications, as the field background now only operates on the whole domain.
Such refactoring is part of this commit.