Fix and modernize the generation scheme for MPI communication tags
The old scheme was causing some tags to be outside of the valid range #982.
Apply fixes and modernization of the tag generation scheme.
Document the requirements of the new scheme.

The remaining pieces of picongpu::CommunicationTag are no longer used for communication.
But they are left to keep the field solver traits, to be fixed in a following PR.