Refactor implementation of fields and field solver traits to not use enum picongpu::CommunicationTag
Those pieces are not connected to communication and used CommunicationTag as a type tag.
Switch the traits to use FieldE and FieldB types instead.
Change the optional parameter of the general GetMargin trait to be a type, not integral constract.
Move specializations in fields solvers from .def files to .hpp files where they belong.
Perform a refactoring of field classes that became possible due to these changes