remove interface `DataConnector::releaseData()`
Remove the interface method `DataConnector::releaseData()`.
The usage and meaning of this function are not clear or defined.
It is not clear how this function should be used in an asynchronous context.
In the past where we introduced this interface, it was designed as the counterpart operation of `DataConnector::get()` but was always unimplemented.