refactore `lockstep::ForEach`
Split the forEach compile time loop into peel and reminder part.
This is slidly better for CPU compiler and is more human readable
compared to the original implementation.
Add two helper types to derive if a functor is callable with or without
Idx as argumenet. This was suggested by @sbastrakov in #3616.