Update summit profile
Use latest openpmd-api/0.13.2 and adios2/2.7.1 module.
To do so, first gcc module hat to be downgraded due to module incompatibilities of gcc/9.3.0 with other required modules.
Second, adios1 can not be used anymore as it is incompatible with e4s module (which is on the other hand required for adios2 and openpmd-api).
But adios1 is deprecated anyway.

LaserWakefield example compiles and runs with these changes on summit.