Carry out invariants in the hot loop of the radiation kernel
There was a factor depending only on frequency and observation direction.
Thus, it didn't change during the hot loop over particles.
Carry this calculation away to constructors of radiation form factors.
Convert retarted time and frequency to float_X to better match the interface.
Add a comment above the hot loop in the radiation kernel.