optimize kernel MoveAndMark
The pusher kernel contains a status flag per supercell which is a marker to know if at least one particle is leaving the supercell.
This status flag is located in shared memory and manipulated via an atomic function by each thread.
This PR aggregates the status per thread first into a register and uses atomic to synchronize the state between threads only if needed and not for each particle that leaves the cell.