Update JUWELS booster profile
The old srun option for memory in getDevice() and getNode() did not work anymore.
Juelich says, the nodes have 512G memory of which *at least* 20G are taken by the system.
Also the partition for getDevice() is changed to `develbooster`, since `gpu` is a different partition with its own profile.
The `booster.tpl` is updated as well.