Add Documentation Page "PIConGPU in 5 Minutes"
The new page in the documentation should provides users with a very
quick overview of how to use PIConGPU and thereby make understanding the
remainder of the documentation more easy, as the user can now understand
the context of the individually described steps.

This draft version lacks cross-references and generally a tight
integration into the existing documentation.