Add missing communication in XrayScattering
I think we discussed when I was writing  the `XrayScattering` plugin that the `GUARD` to `BORDER` communication is not required but I think this was wrong. I finally understood what this does and I think this is missing. Without this the contributions from particles on the edges of the GPUs are not correctly added to the adjacent GPU borders. I was using `Counter` particle shape in my tests back than so this never caused any problems. Note, this is how it is done in other places in `PIConGPU`.