Fix copying of cuSTL device buffers
The implementation was trying to directly write to device addresses from a host side.
It resulted in a segfault with invalid permissions.
This behavior was introduced accidentally in #3862.
That PR kept the wrong of the two pre-existing code paths.
As a result, code was written as if in-kernel, while it is host-only.
Restore the fitting code path.