fix openPMD outputs negative particle patch offset
The patch offset for particles in openPMD output is negate for the top
GPU (y-direction) if sliding window is activated.
In openPMD data this offset is stored as unsigned type and therefor
undowflow and is very large.
A tool which is using particle patches to find domains with particles
within a special range will not work correctly and find all particle
patches.