add experimental PICMI implementation
This commit adds an experimental implementation of the Particle-In-Cell
Modeling Interface (PICMI). Load it with `from picongpu import picmi`.
An example is included in share/picongpu/pypicongpu/examples/warm_plasma

The implementation requires Python 3.9
This has *NOT* been propagated to profiles. However, when loading PICMI
the Python version is checked automatically (and aborted on error).

For further details please refer to the newly added documentation
sections "PICMI" (under Usage) for user documentation and "PyPIConGPU"
for developer documentation and inner workings.

Note that the added documentation sections use diagrams produced by
plantuml. They are *NOT* auto-generated, on update one needs to rebuild
them (run `make`) and re-commit the built pngs. (Rationale: Plantuml is
a quite heavy dependency.)