Add missing double quotes in handleSlurmSignals
Adds missing double quotes around $@ in handleSlurmSignals.sh. According to https://unix.stackexchange.com/a/129077 the difference is that the unquoted expression will treat all arguments in the list (following the expression) as unquoted strings, while '"$@" puts the arguments in double quotes. The unquoted version $@ somehow breaks the picongpu call if any of the picongpu options is passed as a quoted expression, e.g., when passing a json configuration string for openPMD.

Found in an offline discussion with @psychocoderHPC