Fix implementation of pmacc::makeDeepCopy()
It used a wrong constructor of the underlying buffer.
This resulted in effectively making a shallow copy instead of deep copy.
This bug caused incorrect behavior of field background with --duplicateFields enabled since the time it was first introduced.
Clarify the comments for the constructors in question to make such a mistake less likely.