PNG plugin: fix invalid memory access in 2D
The internal variable `sliceDim` is only required to gather information
for 3D simulations.  In 2D the full simulation volume is the slice to
dump. In this case, we need to take care that we do not access extents or
offsets with the array access method of a vector that is outside of
the dimension range.