auto adjust particle exchange buffers for double precision
PIConGPU setups are typically designed for single precision execution.
The memory for particle exchanges defined in `memory.param` is therefore not correct if someone is switching to double precision.

To avoid that users must adjust their setup by hand this PR is automatically adjusting the required buffer sizes by multiplying the size by a factor of 2.