Separate CFL from condition that particle passes at most 1 cell per time step
Though somewhat similar, these conditions are not actually related.
CFL is a mathematical property of field solver.
The latter condition is only required due to our organization of particle pusher and current deposition,
it is not a principle limitation of the method (but is crucial for PIConGPU).
Clarify output of these checks.