fix using uninitilized variable
There was an uninitialized variable in our `AtomicAllInc` method that forced HIP 5.2 to crash during the runtime. The compiler is handling unused conditional variables somehow so that the kernel is failing, even if in the case where we use the variable conditionally uninitialized there should be no problem.
We use warp broadcast to communicate the value of a variable to other threads. In our case, the root value is initialized but the receivers are using an uninitialized variable, which should be no problem.