PMacc: refactor kernel en-queuing
Before this PR PMacc used cupla to execute the device kernel, this came with the drawback that it was not possible to execute an instance of a kernel. Cupla is default constructing the kernel even if PMacc was already using an instance for the kernel.
Therefore a kernel was not allowed to have data members.

This PR refactore the kernel execution method of PMacc and is forwarding the device kernel instance directly to alpaka.

ATTENTION: The interface is backward compatible and therefore PMacc and PIConGPU require no code changes.

It is now allowed to execute a device lambda as a kernel.