remove plugin: PositionsParticle
The plugin is removed because it was not used anymore. The output is
going directly to stdout and only one particle is supported.

If you need such an plugin please use the OpenPMD IO plugins. The
OpenPMD IO plugin is providing structured output, multi particle support
and many more features.