Workaround attribute issue with BP5
The Radiation plugin sets some attributes quite late. This is normally
supported by the openPMD-api, but can become a bit tricky, when the
attribute is a default attribute with a default value.
The Radiation plugin uncovers a bug in (depending on your point of view)
either openPMD-api or in ADIOS2 with this procedure in the BP5 engine.
Realistically the best fix for this would be to set defaults in the
openPMD-api as late as possible, and not as early as possible.
For now, defining these attributes as early as possible in the Radiation
plugin is a sensible workaround.