run CI down-stream pipes parallel
Due to a bug in the gitlab CI it was in the past not possible to run two down-stream pipes in parallel.
In this case, the status of the CI could be wrong.
This bug is now fixed in gitlab, therefore we can run now PIConGPU and PMacc tests in parallel.