remove buffer `*Intern`
The `*Intern` naming of buffers is a result of the very beginning of PIConGPU and was a try to separate the interfaces for buffers a user is allowed to use and what PMacc can be used internally. In old times a user never created a Host or Device buffer, instead, always the GridBuffer was used. Over time we saw that this idea to hide buffer creation was wrong but never found time to refactor this part of the code.