fix incident field calculation order
The incident field is updating the electric field for the CORE and BORDER in different orders.
By moving the electric field CORE update after the incident field update the calculation order for CORE and BORDER is equal.
The original code should not be wrong because updating the electric field is using an addition and this operation is commutative.