avoid that numerical instabilities trigegring the KHI
related to #3234

We observe since long time that our KHi is seeded by nummerical issues
instead a well reproducable condition.
I came up with the idea to use the composit pusher to create maxwell
conform fields over 500 steps and then start to real KHI simulation.
@PrometheusPi suggested to use a random position for electrons and ions
and not using the temperature seeding for electrons.