PMacc: DataConnector change `get()` interface
Change the DataConnector interface to get access to a data set. Remove the parameter to synchronize the data. Synchronization must be called explicit by the user, now.
Before the parameter name was a negation of the execution `noSync` which always required thinking and was error-prone.
The default method used in PMacc and PIConGPU is not performing the synchronization which is now always the case.