refactor `pic-create`
Update our tooling to reflect common workflows.
The python tools will be used mostly from PIConGPU source code and not the input set because we set the python path to PIConGPU source lib folder.
With a new environment variable `PIC_SYSTEM_TEMPLATE_PATH` we can control that only the required profiles and templates for tbg will be compiled by `pic-create`. If the environment variable does not exist we will fall back to the old behavior and copy all system templates/profiles into the input set.

- do not copy picongpu/lib into the input set
- add `PIC_SYSTEM_TEMPLATE_PATH` to control which tbg templates/profiles will be copied