fix crash on AMD (frontier)
fix #4561

Most likly the compiler is removing the floating point rounding guard in
the pusher and therefore the particle position is after the push not in
the valid range `[0;1)`.
Validating the range via a a printf if showing no issue, therefore the
possibility of an compiler bug is very high.

The current workaround add `volatile` to break the compiler
optimization.