fix 2D current deposition performance
EZ and Esirkepov current deposition used the cached/jit shape type of the outer loop for the inner loop.
This results in lower performance because ethe Cached shape which is the default for the outer loop requires more calculations and results in more thread branch differences.

This PR uses the correct shape type for the inner (fast indexing) loop.