collision: reduce memory fragmentation
Reduce the memory heap fragmentation by allocating always a multiple of the typical number of particles per cell for the particle cell list on GPUs.
This change will waste a little bit of memory but on the other side avoid the mallocMC heap fragmentation becoming too large which can result in out-of-memory errors.