fix openPMD plugin output
fix #4641

bug introduced with #4585

In the openPMD plugin we use a param class `ThreadParams` where we
collect many information for the dumping process to avoid forwarding all
in via function parameters. The object is very dirty and we can not
track life-times and different kind of data are stored together.
The bug was introduced because the timestep which was not default
initilized on all systems was required in one of the dumping function
`intersectRangeWithWindow()`  to
calculate the intersection of the range parameter in the openPMD plugin
and the moving window of the simulation.

This PR refactore the interfaces and is passing now the timestep as
function parameter.