fix `setValue`
`setValue()` is not supporintg non trivially copyable value if these
where <=128byte.
For small values we used a optimized version and pass a copy of the
value as parameter to the set kernel, therefore the requirement that the
value type must be trivially copyable is coming from.In case a value
type is not trivially copyable we can us ethe non optimized code path
where we must allocate memory to hold one elment of value type in the
devices main memory.

This PR is changing that the optimzed code path is only used if the
value size if <=128byte and the type is trivially copyable.