fix running muti binary within one MPI context
In case the user is compiling different binaries for different compute nodes or devices and each binary is compiled by other compiler vendors e.g clang, gcc, ... executing all binaries within one MPI context was not possible.
The reason is that tags used for MPI communication were derived from the type signature. Since this approach depends on the order the compiler is creating types in the internal memory the tags used for MPI communication can differ based on the compiler used to generate the binary.

This PR is changing the way PMacc and PICoNGPU are deriving the tag ID for communication.
The id is now depending on the creation order of objects at runtime.