refactor particle init kernel
Motivation: The original implementation assumes that the number of slots in a frame and the number of cells within a supercell are equal. All particles from cell N were always created in the frame slot N too. In case there is a density gradient within the supercell frames are not filled fully which increases the memory footprint until the kernel fill gaps are called.

This PR decuples the number of cells in a supercell and the species frame size. Each thread searches a free slot within the frame and creates the particle in the slot, there is no implicit relation between the slot id in a frame and the cell where the particle is produced in. If the particle creation kernel is called multiple times for the same species the last frame will be filled first if possible.