fix derived field dataset naming
fix #4698

With #4606 #4607 and #4680 we allow that more than one direction of derived fields can be printed. W used `/` to separate the component name from the data/mesh name. This is not valid openPMD or can at least confuse users because `/` is the separator used for openPMD ADIOS data.

This PR is changing the separation to `-` (minus) to not collide with the `_` (underscore) used for the data set name.
Derived fields will be independent data sets even if both show the same property but a different component e.g. x,y,z