remove accessor from `math::Vector<>`
Our implementation of `math::Vector` supports only accessors that provide references, other is not supported by design. To support an accessor that is not returning a reference we need to take care in all implementations if we only read the value or we like to mutate the vector.

Since a generic accessor was not supported and we do not use this feature we will remove the accessor to simplify the implementation.