optimize particle restart
Currently we need to hold all particles for the device in CPU memory to
restart from a checkpoint. On systems with large on device memory but
low amount of CPU memory e.g. ORNL Frontier we can not restart large
simulations.
The OpenPMD plugin has since years chunked particle restart but applied
this chunked load only to the frame de-serialization kernel and not to
the file load.

This PR is loading particles in chunks/batches with openPMD-api, upload
these to the device and is repeating this until all particles are
loaded.
This will reduce the CPU memory footprint to `2 * numParticlePerChunk *
sizeof(Particle)`