remove atomic physics
The atomic physics code in its current form is not to be used and will be replaced with a new version. To avoid that we bring it in under the current work title name `AtomicPhysics2` the current code and documentation are removed.
Remove atomic physics except for species attributes `atomicPhysicsSolver` and `atomicConfigNumber`.
Charge manipulation methods will be kept too.