fix SynchrotronRadiation
Fix  that the current dev is always failing with
```
terminate called after throwing an instance of 'char const*'
```
due to wrong usage of `throw`.
The second issue is that the stage is always consuming memory and is
performing initializing helper fields based on the default param input.
This is one reason why the first issue was showing even if the
synchrotron extension is not activated.