Increase performance of DispersivePulse
Shorten initialization time by reducing the number of frequencies being used in the discrete Fourier transform of the pulse from frequency-space domain to time-space domain.
Usually, not all frequencies in the range Omega=0 to pi/dt will contribute to the field since the laser's spectrum is centered narrowly around the central laser frequency.
Thus, the DFT is limited to the frequencies within +/- 4*sigma around the central laser frequency.