plugin: phaseSpace add pair implementation
`std::pair` is not guaranteed to be trivially copyable therefore we can not pass an object storing a `std::pair` as a member into the device kernel. Depending on the std implementation alpaka will throw a compiletime exception.

By simply implementing a basic pair struct compile, the issue can be avoided.