change atomic physics particle attributes
Atomic physics particle attributes was set by default to a possible
valid value but in case an algorithm forget to set the value correctly
we got a false valid value.
By setting it to an invalid value such errors should become visible at
leat by a crash of the simulation.
An additional advantage is that we can now give an electron some of the
properties which allows to track from which ion state and which
transition state the electron was created.