refactor `value_identifier`
This PR provide the possibility that a value_identifier can fully
describe the operation performed for the
  - first initialization -> `initValue()`
  - if a particle attribute is copied -> `copyValue()`
  - if a particle attribute is derived -> `deriveValue()`

For PIConGPU the binary manipulator `Assign` is renamed into `Derive`
and a `Copy` manipulator is added.

With the new description of value identifier there is no need that we
handle special attributes e.g. `particleId` explicit in any algorihm
where we derive particles.